Changed Language
Changed the developmant from React native to Android Studio (Java) since none of the pedometer assets from React Native worked since 2019.
Currently the app run well in a physical device (It is not able to measure steps with the android virtual device).
The app shows in the screen the total ammount of steps the user has taken and below a step counter which rests to 0 each time the app starts or restarts.

------
Future Plans
------
1.-Delete the total step counter
2.-Make a local DB which will record steps taken in the last 10min, date of registry and the ammount of steps taken in that time.
3.-Export the DB data to a local txt file with the help of a button in the main screen.
4.-Use the DB data to graph the steps taken each day.
5.-Adapt the graph so it looks similar to the reference.
6.- Show Max, Min, Average, and (Median, half and Mode) in the graph screen.